feat(trading-signals): add getTypicalPrice helper
(high + low + close) / 3 was written out three times, in CCI, MFI and
VWAP. All three now share one definition, exported from the package
root so consumers can use it too.

A pure function rather than an indicator: MFI maps its whole candle
window per tick and VWAP recomputes the value for a cached past candle,
neither of which fits a stateful sequential indicator.

Verified against the tulipindicators typprice vectors.
