+++ packages/trading-signals/src/trend/index.ts
@@ -20,7 +20,6 @@ export * from './SWING_HIGH/SwingHigh.js';
 export * from './SWING_LOW/SwingLookback.js';
 export * from './SWING_LOW/SwingLow.js';
 export * from './TEMA/TEMA.js';
-export * from './TYPPRICE/TypicalPrice.js';
 export * from './VSTOP/VolatilityStop.js';
 export * from './VWAP/VWAP.js';
 export * from './WMA/WMA.js';
